+++ nodeconductor/users/views.py
@@ -86,6 +86,11 @@ class InvitationViewSet(ProtectedViewSet):
 
     @detail_route(methods=['post'], filter_backends=[], permission_classes=[permissions.IsAuthenticated])
     def accept(self, request, uuid=None):
+        """ Accept invitation for current user.
+
+            To replace user's email with email from invitation - add parameter
+            'replace_email' to request POST body.
+        """
         invitation = self.get_object()
 
         if invitation.state != models.Invitation.State.PENDING:
@@ -99,7 +104,8 @@ class InvitationViewSet(ProtectedViewSet):
         elif invitation.customer.has_user(request.user):
             raise ValidationError(_('User already has role within this customer.'))
 
-        invitation.accept(request.user)
+        replace_email = bool(request.data.get('replace_email'))
+        invitation.accept(request.user, replace_email=replace_email)
         return Response({'detail': _('Invitation has been successfully accepted.')},
                         status=status.HTTP_200_OK)