+++ CHANGELOG.rst
@@ -5,6 +5,13 @@ Release NEXT
 ------------
 - Nothing yet.
 
+Release 0.133.0
+---------------
+- Remove inner HEAD requests from counter views.
+- Update start time of virtual machine if runtime state changed.
+- Drop VirtualMachineMixin.
+- Allow to define user email on invitation accept.
+
 Release 0.132.0
 ---------------
 - Integrate JSONEditor widgets from django-jsoneditor for JSON fields in admin.